+++ R/dev/devel_script.R
@@ -11,7 +11,6 @@ library(ggforce)
 devtools::load_all("./")
 source("R/dev/10_functions_plot.R")
 
-# Prepare data for plotting
 p <- plotRiparian(hdf5_fn = "~/workspace/orthology/output/benchmark/orthopair/hdf5_out/Osat_Ogla.h5",
                   genomes = c("Osat", "Ogla"), 
                   select_chr = list(Osat = 1:12, Ogla = 1:12),
@@ -31,17 +30,37 @@ p <- plotRiparian(hdf5_fn = "~/workspace/orthology/output/benchmark/orthopair/hd
                   normalize_genome = TRUE,
                   genome_scale = 100)
 
-ggsave("~/workspace/orthology/output/benchmark/synteny_stacked.pdf",
+ggsave("~/workspace/orthology/output/benchmark/synteny_stacked_min1e5_rice.pdf",
        p, width = 8, height = 4)
 
-
-
-# Prepare data for plotting
+# Plant kingdom wide
 p <- plotRiparian(hdf5_fn = "~/workspace/orthology/output/benchmark_plant/orthopair/reorg_out/reorg_orthopair.h5",
                   genomes = c("Osat", "Hvul", "Atha", "Mpol"),
                   select_chr = list(Osat = 1:12, Hvul = 1:7, Atha = 1:5, Mpol = 1:8),
                   chr_sizes = NULL,
-                  chr_gap = 3,
+                  chr_gap = 6,
+                  track_gap = 0.5,
+                  max_links = 50000,
+                  seed = 1,
+                  alpha_base = NULL,
+                  linewidth = 0.45,
+                  inv_darken = 0.35,
+                  inv_alpha_mult = 1.6,
+                  palette = c("#d73027","#fc8d59","#fee08b","#91bfdb","#4575b4"),
+                  min_block_width = 0,
+                  chr_bar_lw = 8,
+                  chr_label_size = 15,
+                  normalize_genome = TRUE,
+                  genome_scale = 200)
+
+ggsave("~/workspace/orthology/output/benchmark_plant/synteny_stacked_min0_plant.pdf",
+       p, width = 8, height = 8)
+
+p <- plotRiparian(hdf5_fn = "~/workspace/orthology/output/benchmark_plant/orthopair/reorg_out/reorg_orthopair.h5",
+                  genomes = c("Osat", "Hvul", "Atha", "Mpol"),
+                  select_chr = list(Osat = 1:12, Hvul = 1:7, Atha = 1:5, Mpol = 1:8),
+                  chr_sizes = NULL,
+                  chr_gap = 6,
                   track_gap = 0.5,
                   max_links = 50000,
                   seed = 1,
@@ -56,5 +75,5 @@ p <- plotRiparian(hdf5_fn = "~/workspace/orthology/output/benchmark_plant/orthop
                   normalize_genome = TRUE,
                   genome_scale = 200)
 
-ggsave("~/workspace/orthology/output/benchmark_plant/synteny_stacked.pdf",
+ggsave("~/workspace/orthology/output/benchmark_plant/synteny_stacked_min1e5_plant.pdf",
        p, width = 8, height = 8)